+++ py_trees/composites.py
@@ -35,8 +35,6 @@ Perform the checks or actions you need to do in the non-composite behaviours.
 # Imports
 ##############################################################################
 
-
-
 import itertools
 
 from . import common
@@ -62,7 +60,8 @@ class Composite(Behaviour):
     def __init__(self, name="", children=None, *args, **kwargs):
         super(Composite, self).__init__(name, *args, **kwargs)
         if children is not None:
-            list(map(self.add_child, children))
+            for child in children:
+                self.add_child(child)
         else:
             self.children = []