+++ py_trees/demos/context_switching.py
@@ -137,7 +137,7 @@ def main():
     Entry point for the demo script.
     """
     args = command_line_argument_parser().parse_args()
-    print((description()))
+    print(description())
     py_trees.logging.level = py_trees.logging.Level.DEBUG
 
     tree = create_tree()
@@ -155,7 +155,7 @@ def main():
     tree.setup(timeout=15)
     for i in range(1, 6):
         try:
-            print(("\n--------- Tick {0} ---------\n".format(i)))
+            print("\n--------- Tick {0} ---------\n".format(i))
             tree.tick_once()
             print("\n")
             py_trees.display.print_ascii_tree(tree, show_status=True)