+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Forthcoming
+-----------
+* Python 2/3 compatibility
+
 0.5.10 (2017-06-17)
 -------------------
 * [meta] add children monkeypatching for composite imposters